Automatically backed up by Learn

+++ db/migrate/002_create_costume_stores.rb
@@ -4,7 +4,7 @@ class CreateCostumeStores < ActiveRecord::Migration[5.1]
       t.string :name
       t.string :location
       t.float :costume_inventory
-      t.float :num_of_employees
+      t.integer :num_of_employees
       t.boolean :still_in_business
       t.float :opening_time 
       t.float :closing_time
